Added auth solved some bugs
